task of the week

+++ shOOP/Client.cpp
@@ -1,7 +1,5 @@
 #include "stdafx.h"
 #include "Client.h"
-#include <iostream>
-#include <cstdlib>
 
 
 Client::Client()
@@ -50,7 +48,7 @@ std::string Client::serialize()
 	temp += EIK;
 	temp += ", IN: ";
 	temp += IN;
-
+	
 	serialized = temp;
 
 	return temp;
@@ -91,7 +89,7 @@ Client* Client::deserialize(std::string serialized)
 			temp += serialized[i];
 			i++;
 		}
-		newClient->type = std::atoll(temp.c_str());
+		newClient->type = std::atoi(temp.c_str());
 		temp = "";
 		while (serialized[i] != ':' && i < len) i++;
 		i += 2;
